+++ command/genesis/params.go
@@ -3,16 +3,13 @@ package genesis
 import (
 	"errors"
 	"fmt"
-	"math/big"
 	"os"
-	"strings"
 	"time"
 
 	"github.com/0xPolygon/polygon-edge/chain"
 	"github.com/0xPolygon/polygon-edge/command"
 	"github.com/0xPolygon/polygon-edge/command/helper"
 	"github.com/0xPolygon/polygon-edge/consensus/polybft"
-	"github.com/0xPolygon/polygon-edge/contracts"
 	"github.com/0xPolygon/polygon-edge/server"
 	"github.com/0xPolygon/polygon-edge/types"
 )
@@ -32,11 +29,6 @@ const (
 	rewardWalletFlag             = "reward-wallet"
 	blockTrackerPollIntervalFlag = "block-tracker-poll-interval"
 	proxyContractsAdminFlag      = "proxy-contracts-admin"
-)
-
-// Legacy flags that need to be preserved for running clients
-const (
-	chainIDFlagLEGACY = "chainid"
 )
 
 var (
@@ -82,7 +74,6 @@ type genesisParams struct {
 	validators           []string
 
 	extraData []byte
-	consensus server.ConsensusType
 
 	consensusEngineConfig map[string]interface{}
 
@@ -149,31 +140,8 @@ func (p *genesisParams) validateFlags() error {
 	}
 
 	if p.isPolyBFTConsensus() {
-		if err := p.extractNativeTokenMetadata(); err != nil {
+		if err := p.validatePolyBFT(); err != nil {
 			return err
-		}
-
-		if err := p.validateBurnContract(); err != nil {
-			return err
-		}
-
-		if err := p.validateRewardWalletAndToken(); err != nil {
-			return err
-		}
-
-		if err := p.validatePremineInfo(); err != nil {
-			return err
-		}
-
-		if err := p.validateProxyContractsAdmin(); err != nil {
-			return err
-		}
-
-		if p.epochSize < 2 {
-			// Epoch size must be greater than 1, so new transactions have a chance to be added to a block.
-			// Otherwise, every block would be an endblock (meaning it will not have any transactions).
-			// Check is placed here to avoid additional parsing if epochSize < 2
-			return errInvalidEpochSize
 		}
 	}
 
@@ -266,33 +234,6 @@ func (p *genesisParams) initGenesisConfig() error {
 	return nil
 }
 
-// validateRewardWalletAndToken validates reward wallet flag
-func (p *genesisParams) validateRewardWalletAndToken() error {
-	if p.rewardWallet == "" {
-		return errRewardWalletNotDefined
-	}
-
-	if !p.nativeTokenConfig.IsMintable && p.rewardTokenCode == "" {
-		return errRewardTokenOnNonMintable
-	}
-
-	premineInfo, err := helper.ParsePremineInfo(p.rewardWallet)
-	if err != nil {
-		return err
-	}
-
-	if premineInfo.Address == types.ZeroAddress {
-		return errRewardWalletZero
-	}
-
-	// If epoch rewards are enabled, reward wallet must have some amount of premine
-	if p.epochReward > 0 && premineInfo.Amount.Cmp(big.NewInt(0)) < 1 {
-		return errRewardWalletAmountZero
-	}
-
-	return nil
-}
-
 // parsePremineInfo parses premine flag
 func (p *genesisParams) parsePremineInfo() error {
 	p.premineInfos = make([]*helper.PremineInfo, 0, len(p.premine))
@@ -304,52 +245,6 @@ func (p *genesisParams) parsePremineInfo() error {
 		}
 
 		p.premineInfos = append(p.premineInfos, premineInfo)
-	}
-
-	return nil
-}
-
-// validatePremineInfo validates whether reserve account (0x0 address) is premined
-func (p *genesisParams) validatePremineInfo() error {
-	for _, premineInfo := range p.premineInfos {
-		if premineInfo.Address == types.ZeroAddress {
-			// we have premine of zero address, just return
-			return nil
-		}
-	}
-
-	return errReserveAccMustBePremined
-}
-
-// validateBlockTrackerPollInterval validates block tracker block interval
-// which can not be 0
-func (p *genesisParams) validateBlockTrackerPollInterval() error {
-	if p.blockTrackerPollInterval == 0 {
-		return helper.ErrBlockTrackerPollInterval
-	}
-
-	return nil
-}
-
-// validateBurnContract validates burn contract. If native token is mintable,
-// burn contract flag must not be set. If native token is non mintable only one burn contract
-// can be set and the specified address will be used to predeploy default EIP1559 burn contract.
-func (p *genesisParams) validateBurnContract() error {
-	if p.isBurnContractEnabled() {
-		burnContractInfo, err := parseBurnContractInfo(p.burnContract)
-		if err != nil {
-			return fmt.Errorf("invalid burn contract info provided: %w", err)
-		}
-
-		if p.nativeTokenConfig.IsMintable {
-			if burnContractInfo.Address != types.ZeroAddress {
-				return errors.New("only zero address is allowed as burn destination for mintable native token")
-			}
-		} else {
-			if burnContractInfo.Address == types.ZeroAddress {
-				return errors.New("it is not allowed to deploy burn contract to 0x0 address")
-			}
-		}
 	}
 
 	return nil
@@ -382,38 +277,9 @@ func (p *genesisParams) validateGenesisBaseFeeConfig() error {
 	return nil
 }
 
-func (p *genesisParams) validateProxyContractsAdmin() error {
-	if strings.TrimSpace(p.proxyContractsAdmin) == "" {
-		return errors.New("proxy contracts admin address must be set")
-	}
-
-	proxyContractsAdminAddr := types.StringToAddress(p.proxyContractsAdmin)
-	if proxyContractsAdminAddr == types.ZeroAddress {
-		return errors.New("proxy contracts admin address must not be zero address")
-	}
-
-	if proxyContractsAdminAddr == contracts.SystemCaller {
-		return errors.New("proxy contracts admin address must not be system caller address")
-	}
-
-	return nil
-}
-
 // isBurnContractEnabled returns true in case burn contract info is provided
 func (p *genesisParams) isBurnContractEnabled() bool {
 	return p.burnContract != ""
-}
-
-// extractNativeTokenMetadata parses provided native token metadata (such as name, symbol and decimals count)
-func (p *genesisParams) extractNativeTokenMetadata() error {
-	tokenConfig, err := polybft.ParseRawTokenConfig(p.nativeTokenConfigRaw)
-	if err != nil {
-		return err
-	}
-
-	p.nativeTokenConfig = tokenConfig
-
-	return nil
 }
 
 func (p *genesisParams) getResult() command.CommandResult {
