--- e2e/const.go
@@ -1,12 +1,0 @@
-//nolint:lll
-package e2e
-
-// Data for mock contract
-const sampleByteCode = `608060405234801561001057600080fd5b50610286806100206000396000f3fe6080604052600436106100345760003560e01c8063498e6857146100395780636e7e996e14610043578063dd13c6171461004d575b600080fd5b610041610069565b005b61004b6100da565b005b610067600480360361006291908101906101be565b61014b565b005b600073ffffffffffffffffffffffffffffffffffffffff1673010000000000000000000000000000000000000073ffffffffffffffffffffffffffffffffffffffff167fec89f5028137f6e2218f9178e2ddfa454a509f4778d9cf323e96f42a902d307f60405160405180910390a3565b73010000000000000000000000000000000000000073ffffffffffffffffffffffffffffffffffffffff16600073ffffffffffffffffffffffffffffffffffffffff167fec89f5028137f6e2218f9178e2ddfa454a509f4778d9cf323e96f42a902d307f60405160405180910390a3565b8073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff167fec89f5028137f6e2218f9178e2ddfa454a509f4778d9cf323e96f42a902d307f60405160405180910390a35050565b6000813590506101b88161022c565b92915050565b600080604083850312156101d157600080fd5b60006101df858286016101a9565b92505060206101f0858286016101a9565b9150509250929050565b60006102058261020c565b9050919050565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b610235816101fa565b811461024057600080fd5b5056fea365627a7a72315820874667b781b97cc25ada056a6b1d529cadb5d6e734754990b055e2d072d506d16c6578706572696d656e74616cf564736f6c63430005110040`
-const bloomFilterTestBytecode = `608060405234801561001057600080fd5b5060fa8061001f6000396000f3fe6080604052348015600f57600080fd5b506004361060285760003560e01c806362b9ab5b14602d575b600080fd5b60336035565b005b3373ffffffffffffffffffffffffffffffffffffffff167fdf50c7bb3b25f812aedef81bc334454040e7b27e27de95a79451d663013b7e17602a604051607a91906091565b60405180910390a2565b608b8160b4565b82525050565b600060208201905060a460008301846084565b92915050565b6000819050919050565b600060bd8260aa565b905091905056fea2646970667358221220052488d0b0bc3f50f2d498d982ba7d31d724617d6055258426a7b819be800bf864736f6c63430008000033`
-
-// Default settings for IBFT nodes
-const (
-	IBFTDirPrefix = "e2e-ibft-"
-	IBFTMinNodes  = 4
-)
