+++ scripts/README.md
@@ -12,7 +12,6 @@ When deploying with `polybft` consensus, there are some additional dependencies:
 
 Running `polygon-edge` from local binary can be done very easily by using provided `scripts` folder.
 
-* `scripts/cluster ibft` - deploy environment with `ibft` consensus
 * `scripts/cluster polybft` - deploy environment with `polybft` consensus
 
 ## Customization
