+++ types/header_hash.go
@@ -8,8 +8,7 @@ import (
 var HeaderHash func(h *Header) Hash
 
 // This is the default header hash for the block.
-// In IBFT, this header hash method is substituted
-// for Istanbul Header Hash calculation
+// In PolyBFT, this header hash method is substituted with different Header Hash calculation
 func init() {
 	HeaderHash = defHeaderHash
 }
